Only remove file if it exists.

+++ Makefile
@@ -48,7 +48,7 @@ check-unbind-helper:
 	@[ "$$(stat -c "%a %U" bin/unbind-helper)" = "4755 root" ]
 
 clean-unbind-helper:
-	sudo rm bin/unbind-helper
+	if [ -e bin/unbind-helper ]; then sudo rm bin/unbind-helper; fi
 
 # udev rules
 install-udev:
